+++ core-shared-ui/src/main/java/com/codandotv/streamplayerapp/core_shared_ui/utils/Sharing.kt
@@ -12,8 +12,6 @@ object Sharing {
     const val COPY_CONTENT_TYPE_TEXT = "text"
     const val WHATSAPP_PACKAGE_SHARING = "com.whatsapp"
     const val INSTAGRAM_PACKAGE_SHARING = "com.instagram.android"
-    const val INSTAGRAM_IMAGE_PREFIX = "IMG:"
-    const val INSTAGRAM_IMAGE_BOUND = 100000
     const val INSTAGRAM_STORY_DESTINATION = "com.instagram.share.ADD_TO_STORY"
     const val SMS_CONTENT_TYPE = "sms:"
     const val SMS_CONTENT_BODY = "sms_body"
