+++ core-shared-ui/src/main/res/values/strings.xml
@@ -17,6 +17,7 @@
     <string name="sharing_title_more_options">Mais Opções</string>
     <string name="sharing_whatsapp_message">Já assistiu a \"%1s\" na Netflix?\n\n%2s</string>
     <string name="whatsapp_not_installed_message">Por favor, instale o aplicativo WhatsApp</string>
+    <string name="instagram_not_installed_message">Por favor, instale o aplicativo Instagram</string>
     <string name="sms_app_error_message">Erro ao abrir o app Mensagens</string>
     <!-- endregion Sharing Content   -->
 </resources>
